backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
pydantic[email]==2.10.4
pydantic-settings==2.7.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
redis==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
twilio==9.4.3
sendgrid==6.11.0
stripe==11.4.0
anthropic==0.42.0
celery==5.4.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
logfire==2.7.0
weasyprint==64.1
openpyxl==3.1.5
python-dotenv==1.0.1
slowapi==0.1.9
